-Moved Non generic ISpread and IDiffSpread to separate namespace -Fixed nodes using Non generic namespaces
