+++ vvvv45/src/core/Hosting/IO/Pins/Config/ConfigPin.cs
@@ -4,6 +4,7 @@ using VVVV.Hosting.Pins.Input;
 using VVVV.PluginInterfaces.V1;
 using VVVV.PluginInterfaces.V2;
 using VVVV.Utils.Streams;
+using VVVV.PluginInterfaces.V2.NonGeneric;
 
 namespace VVVV.Hosting.Pins.Config
 {
