+++ vvvv45/src/core/Hosting/IO/SpreadRegistry.cs
@@ -6,6 +6,7 @@ using VVVV.Hosting.Pins.Output;
 using VVVV.PluginInterfaces.V1;
 using VVVV.PluginInterfaces.V2;
 using VVVV.Utils.Streams;
+using VVVV.PluginInterfaces.V2.NonGeneric;
 
 namespace VVVV.Hosting.IO
 {
