+++ vvvv45/src/core/PluginInterfaces/PluginInterfaces.csproj
@@ -98,6 +98,7 @@
     <Compile Include="V2\IO\IOBuildContext.cs" />
     <Compile Include="V2\IPluginHost2.cs" />
     <Compile Include="V2\IPlugin2.cs" />
+    <Compile Include="V2\NonGenericISpread.cs" />
     <Compile Include="V2\PluginHostExtensions.cs" />
     <Compile Include="V2\IStartable.cs" />
     <Compile Include="V2\IStartableRegistry.cs" />
