+++ vvvv45/src/core/PluginInterfaces/V2/ISpread.cs
@@ -7,6 +7,7 @@ using System.Runtime.InteropServices;
 using System.Text;
 using VVVV.Utils.Streams;
 using VVVV.Utils.VMath;
+using VVVV.PluginInterfaces.V2.NonGeneric;
 
 namespace VVVV.PluginInterfaces.V2
 {
@@ -14,27 +15,7 @@ namespace VVVV.PluginInterfaces.V2
 	/// Common non-generic interface to the underlying input/output/config pins.
 	/// Set/Get, Read/Write methods are only implemented when it makes sense.
 	/// </summary>
-	[ComVisible(false)]
-	public interface ISpread : IEnumerable, ICloneable, ISynchronizable, IFlushable
-	{
-		/// <summary>
-		/// Provides random read/write access to the actual data.
-		/// </summary>
-		object this[int index]
-		{
-			get;
-			set;
-		}
-		
-		/// <summary>
-		/// Get/Set the size of this spread.
-		/// </summary>
-		int SliceCount
-		{
-			get;
-			set;
-		}
-	}
+
 	
 	/// <summary>
 	/// Common interface to the underlying input/output/config pins.
