+++ vvvv45/src/core/PluginInterfaces/V2/SpreadUtils.cs
@@ -1,5 +1,6 @@
 ﻿using System;
 using System.Diagnostics;
+using VVVV.PluginInterfaces.V2.NonGeneric;
 
 namespace VVVV.PluginInterfaces.V2
 {
