+++ vvvv45/src/nodes/plugins/vvvv/Kontrolleur/KontrolleurNode.cs
@@ -18,6 +18,7 @@ using VVVV.PluginInterfaces.V2.Graph;
 using VVVV.Utils.VColor;
 using VVVV.Utils.VMath;
 using VVVV.Utils.OSC;
+using VVVV.PluginInterfaces.V2.NonGeneric;
 #endregion usings
 
 namespace VVVV.Nodes
